Add ROS wrapper for Moteus
Included in this commit is the outline of the moteus wrapper
class that will be included in the final pen plotter, i.e.
the RaspberryPiMoteusWrapper. The class is intended to act as
a stand-alone object; i.e., creating a single object is sufficient
to bring the Motues motors into ROS.

The Python 'controls' module is erroneously not found by Python in this
commit and does not properly execute. The C++ executable generated
by the package, however, does correctly locate the message headers.